Reland "Upconvert various types to int.", common_audio portion.
This reverts portions of commit cb180976dd0e9672cde4523d87b5f4857478b5e9, which
reverted commit 83ad33a8aed1fb00e422b6abd33c3e8942821c24.  Specifically, the
files in webrtc/common_audio/ are relanded.

The original commit message is below:

Upconvert various types to int.

Per comments from HL/kwiberg on https://webrtc-codereview.appspot.com/42569004 , when there is existing usage of mixed types (int16_t, int, etc.), we'd prefer to standardize on larger types like int and phase out use of int16_t.

Specifically, "Using int16 just because we're sure all reasonable values will fit in 16 bits isn't usually meaningful in C."

This converts some existing uses of int16_t (and, in a few cases, other types such as uint16_t) to int (or, in a few places, int32_t).  Other locations will be converted to size_t in a separate change.

BUG=none
TBR=andrew

Review URL: https://codereview.webrtc.org/1184613003

Cr-Commit-Position: refs/heads/master@{#9425}

+++ webrtc/common_audio/signal_processing/cross_correlation.c
@@ -16,8 +16,8 @@ void WebRtcSpl_CrossCorrelationC(int32_t* cross_correlation,
                                  const int16_t* seq2,
                                  int16_t dim_seq,
                                  int16_t dim_cross_correlation,
-                                 int16_t right_shifts,
-                                 int16_t step_seq2) {
+                                 int right_shifts,
+                                 int step_seq2) {
   int i = 0, j = 0;
 
   for (i = 0; i < dim_cross_correlation; i++) {
